fix: resolve Windows type mismatches
- Change .as_bool() to .is_err() for EscapeCommFunction Result return type
- Wrap raw handle in HANDLE() for windows crate 0.52 compatibility
- Cast PageFaultCount from u32 to u64

+++ src/monitoring/windows.rs
@@ -126,7 +126,7 @@ impl WindowsPerformanceMonitor {
                 metrics.peak_working_set_size = pmc.PeakWorkingSetSize as usize;
                 metrics.page_file_usage = pmc.PagefileUsage as usize;
                 metrics.peak_page_file_usage = pmc.PeakPagefileUsage as usize;
-                metrics.page_fault_count = pmc.PageFaultCount;
+                metrics.page_fault_count = pmc.PageFaultCount as u64;
             }
 
             // Get CPU times and calculate usage
